refactor: arrangements builders
