+++ bin/tests/change_command_displays_and_speakers_on_windows_should.rs
@@ -2,7 +2,7 @@
 
 use convertible_couch_lib::{application_error::ApplicationError, func};
 use convertible_couch_testing::arrangements::{
-    builders::{ApplicationBuilder, ArgumentsBuilder},
+    builders::{application::ApplicationBuilder, arguments::ArgumentsBuilder},
     fuzzing::{ComputerBuilder, Fuzzer},
 };
