+++ bin/tests/change_command_displays_only_should.rs
@@ -5,7 +5,7 @@ use convertible_couch_lib::{
     func,
 };
 use convertible_couch_testing::arrangements::{
-    builders::{ApplicationBuilder, ArgumentsBuilder},
+    builders::{application::ApplicationBuilder, arguments::ArgumentsBuilder},
     fuzzing::{ComputerBuilder, Fuzzer},
 };
