+++ bin/tests/change_command_speakers_only_should.rs
@@ -3,7 +3,7 @@ use convertible_couch_lib::{
     application_error::ApplicationError, func, speakers_settings::SpeakersSettingsResult,
 };
 use convertible_couch_testing::arrangements::{
-    builders::{ApplicationBuilder, ArgumentsBuilder},
+    builders::{application::ApplicationBuilder, arguments::ArgumentsBuilder},
     fuzzing::{ComputerBuilder, Fuzzer},
 };
