+++ testing/src/arrangements/builders/mod.rs
@@ -1,0 +1,3 @@
+pub mod application;
+pub mod arguments;
+pub mod command_result;
